# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "operationId": "WebhooksController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Webhook endpoint retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get the webhook endpoint for the current mode",
        "tags": [
          "Webhooks"
        ]
      },
      "put": {
        "operationId": "WebhooksController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook endpoint saved successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create or update the webhook endpoint for the current mode",
        "tags": [
          "Webhooks"
        ]
      },
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Webhook endpoint deleted successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Delete the webhook endpoint for the current mode",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/v1/webhooks/events": {
      "get": {
        "operationId": "WebhooksController_findEvents",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by delivery status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "delivered",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook events retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List webhook delivery events for the current mode",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/v1/webhooks/events/{id}": {
      "get": {
        "operationId": "WebhooksController_findEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Webhook event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook event retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get a single webhook delivery event",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dexxify API",
    "description": "Crypto Infrastructure API for Africa (Wallets, Payouts, Offramp, Onramp, KYC, KYB)",
    "version": "1.0",
    "contact": {
      "name": "Dexxify",
      "url": "https://www.dexxify.com",
      "email": "dexxifyhq@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://api.dexxify.com/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "API key for /api/v1/* endpoints"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "errors": {
            "type": "object",
            "example": [
              "amount must not be less than 0"
            ]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-17T12:00:00.000Z"
          }
        },
        "required": [
          "success",
          "status",
          "message",
          "timestamp"
        ]
      },
...
      "SaveWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Webhook URL",
            "example": "https://example.com/webhook"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the webhook endpoint is enabled",
            "example": true
          }
        }
      },
...
    }
  }
}